feat: add feature flags
