+++ Cargo.toml
@@ -4,8 +4,6 @@ version = "0.1.0"
 license-file = "LICENSE"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
 async-trait = "0.1"
 aws-config = "0.0.22-alpha"
@@ -15,9 +13,9 @@ aws-smithy-client = { version = "0.27.0-alpha", features = ["test-util"] }
 aws-smithy-http = "0.27.0-alpha"
 aws-types = "0.0.22-alpha"
 futures = { version = "0.3", features = ["std"] }
-lambda_runtime = "0.4"
-lambda_http = "0.4"
-rayon = "1.5"
+lambda_runtime = { version = "0.4", optional = true }
+lambda_http = { version = "0.4", optional = true }
+rayon = { version = "1.5", optional = true }
 serde = "1"
 serde_json = "1.0"
 tracing = "0.1"
@@ -29,3 +27,37 @@ float-cmp = "0.9"
 http = "0.2"
 rand = "0.8"
 reqwest = { version = "0.11", features = ["json"] }
+
+[features]
+default = ["lambda"]
+lambda = ["lambda_runtime", "lambda_http", "rayon"]
+
+[[bin]]
+name = "delete-product"
+path = "src/bin/lambda/delete-product.rs"
+test = false
+required-features = ["lambda"]
+
+[[bin]]
+name = "get-product"
+path = "src/bin/lambda/get-product.rs"
+test = false
+required-features = ["lambda"]
+
+[[bin]]
+name = "get-products"
+path = "src/bin/lambda/get-products.rs"
+test = false
+required-features = ["lambda"]
+
+[[bin]]
+name = "put-product"
+path = "src/bin/lambda/put-product.rs"
+test = false
+required-features = ["lambda"]
+
+[[bin]]
+name = "dynamodb-streams"
+path = "src/bin/lambda/dynamodb-streams.rs"
+test = false
+required-features = ["lambda"]
