+++ src/bin/lambda/delete-product.rs
@@ -3,12 +3,10 @@ use lambda_http::{
     lambda_runtime::{self, Context},
     Request,
 };
-use products::{entrypoints::lambda_apigateway::get_products, utils::*};
-
-type E = Box<dyn std::error::Error + Send + Sync + 'static>;
+use products::{entrypoints::lambda::apigateway::delete_product, utils::*};
 
 #[tokio::main]
-async fn main() -> Result<(), E> {
+async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
     // Initialize logger
     setup_tracing();
 
@@ -19,19 +17,19 @@ async fn main() -> Result<(), E> {
     //
     // This is the entry point for the Lambda function. The `lambda_runtime`
     // crate will take care of contacting the Lambda runtime API and invoking
-    // the `get_products` function.
+    // the `delete_product` function.
     // See https://docs.aws.amazon.com/lambda/latest/dg/runtimes-api.html
     //
     // This uses a closure to pass the Service without having to reinstantiate
     // it for every call. This is a bit of a hack, but it's the only way to
     // pass a service to a lambda function.
     //
-    // Furthermore, we don't await the result of `get_products` because
+    // Furthermore, we don't await the result of `delete_product` because
     // async closures aren't stable yet. This way, the closure returns a Future,
     // which matches the signature of the lambda function.
     // See https://github.com/rust-lang/rust/issues/62290
     lambda_runtime::run(handler(|event: Request, ctx: Context| {
-        get_products(&service, event, ctx)
+        delete_product(&service, event, ctx)
     }))
     .await?;
     Ok(())
