+++ src/bin/lambda/get-product.rs
@@ -1,11 +1,14 @@
-use lambda_runtime::{handler_fn, Context};
-use products::{
-    entrypoints::lambda_dynamodb::{model::DynamoDBEvent, parse_events},
-    utils::*,
+use lambda_http::{
+    handler,
+    lambda_runtime::{self, Context},
+    Request,
 };
+use products::{entrypoints::lambda::apigateway::get_product, utils::*};
+
+type E = Box<dyn std::error::Error + Send + Sync + 'static>;
 
 #[tokio::main]
-async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
+async fn main() -> Result<(), E> {
     // Initialize logger
     setup_tracing();
 
@@ -16,19 +19,19 @@ async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>
     //
     // This is the entry point for the Lambda function. The `lambda_runtime`
     // crate will take care of contacting the Lambda runtime API and invoking
-    // the `delete_product` function.
+    // the `get_product` function.
     // See https://docs.aws.amazon.com/lambda/latest/dg/runtimes-api.html
     //
     // This uses a closure to pass the Service without having to reinstantiate
     // it for every call. This is a bit of a hack, but it's the only way to
     // pass a service to a lambda function.
     //
-    // Furthermore, we don't await the result of `delete_product` because
+    // Furthermore, we don't await the result of `get_product` because
     // async closures aren't stable yet. This way, the closure returns a Future,
     // which matches the signature of the lambda function.
     // See https://github.com/rust-lang/rust/issues/62290
-    lambda_runtime::run(handler_fn(|event: DynamoDBEvent, ctx: Context| {
-        parse_events(&service, event, ctx)
+    lambda_runtime::run(handler(|event: Request, ctx: Context| {
+        get_product(&service, event, ctx)
     }))
     .await?;
     Ok(())
