+++ src/bin/lambda/get-products.rs
@@ -3,10 +3,12 @@ use lambda_http::{
     lambda_runtime::{self, Context},
     Request,
 };
-use products::{entrypoints::lambda_apigateway::delete_product, utils::*};
+use products::{entrypoints::lambda::apigateway::get_products, utils::*};
+
+type E = Box<dyn std::error::Error + Send + Sync + 'static>;
 
 #[tokio::main]
-async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
+async fn main() -> Result<(), E> {
     // Initialize logger
     setup_tracing();
 
@@ -17,19 +19,19 @@ async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>
     //
     // This is the entry point for the Lambda function. The `lambda_runtime`
     // crate will take care of contacting the Lambda runtime API and invoking
-    // the `delete_product` function.
+    // the `get_products` function.
     // See https://docs.aws.amazon.com/lambda/latest/dg/runtimes-api.html
     //
     // This uses a closure to pass the Service without having to reinstantiate
     // it for every call. This is a bit of a hack, but it's the only way to
     // pass a service to a lambda function.
     //
-    // Furthermore, we don't await the result of `delete_product` because
+    // Furthermore, we don't await the result of `get_products` because
     // async closures aren't stable yet. This way, the closure returns a Future,
     // which matches the signature of the lambda function.
     // See https://github.com/rust-lang/rust/issues/62290
     lambda_runtime::run(handler(|event: Request, ctx: Context| {
-        delete_product(&service, event, ctx)
+        get_products(&service, event, ctx)
     }))
     .await?;
     Ok(())
