+++ src/entrypoints/lambda/dynamodb/mod.rs
