+++ src/entrypoints/lambda/dynamodb/model.rs
