+++ src/entrypoints/lambda/mod.rs
@@ -1,0 +1,2 @@
+pub mod apigateway;
+pub mod dynamodb;
